requirements.txt
requests
simplejson
binance-futures-connector
